+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 == 1.1 (May 23, 2014)
 
-* Overall development clean-up
 * Fix error thrown in certain cases when the editor was loosing focus
+* Move the editor toolbar in the body, so we can use `overflow: hidden` in containers
+* Improve the editor toolbar positioning
+* Overall development environment clean-up
 
 == 1.0.0 (November 11, 2013)